allow overriding the max latency of the filter
This can be temporarily used at runtime to synchronize a low-latency
filter (i.e. dead reckoning) with a high-latency one (i.e. global
positioning) by feeding the latter in the xy_samples port of the
former. Obviously, it must be done in a (control) state where the
fast latency filter is unused.